Rename and promote nested class
